+++ src/XrdPosix/XrdPosixExtern.hh
@@ -105,8 +105,7 @@ extern int        XrdPosix_Fdatasync(int fildes);
 extern int        XrdPosix_Fflush(FILE *stream);
 
 #if defined(__linux__) || defined(__GNU__) || (defined(__FreeBSD_kernel__) && defined(__GLIBC__))
-extern long long  XrdPosix_Fgetxattr (int fd, const char *name,
-                                      void *value, unsigned long long size);
+extern ssize_t    XrdPosix_Fgetxattr (int fd, const char *name, void *value, size_t size);
 #endif
 
 extern FILE      *XrdPosix_Fopen(const char *path, const char *mode);
@@ -115,7 +114,7 @@ extern size_t     XrdPosix_Fread(void *ptr, size_t size, size_t nitems, FILE *st
 
 extern int        XrdPosix_Fseek(FILE *stream, long offset, int whence);
 
-extern int        XrdPosix_Fseeko(FILE *stream, off64_t offset, int whence);
+extern int        XrdPosix_Fseeko(FILE *stream, off_t offset, int whence);
 
 extern int        XrdPosix_Fstat(int fildes, struct stat *buf);
 
@@ -127,21 +126,19 @@ extern int        XrdPosix_Fsync(int fildes);
 
 extern long       XrdPosix_Ftell(FILE *stream);
 
-extern off64_t  XrdPosix_Ftello(FILE *stream);
+extern off_t      XrdPosix_Ftello(FILE *stream);
 
-extern int        XrdPosix_Ftruncate(int fildes, off64_t offset);
+extern int        XrdPosix_Ftruncate(int fildes, off_t offset);
 
 extern size_t     XrdPosix_Fwrite(const void *ptr, size_t size, size_t nitems, FILE *stream);
 
 #if defined(__linux__) || defined(__GNU__) || (defined(__FreeBSD_kernel__) && defined(__GLIBC__))
-extern off64_t  XrdPosix_Getxattr (const char *path, const char *name,
-                                     void *value, unsigned long long size);
+extern ssize_t    XrdPosix_Getxattr (const char *path, const char *name, void *value, size_t size);
 
-extern off64_t  XrdPosix_Lgetxattr(const char *path, const char *name,
-                                     void *value, unsigned long long size);
+extern ssize_t    XrdPosix_Lgetxattr(const char *path, const char *name, void *value, size_t size);
 #endif
 
-extern off64_t  XrdPosix_Lseek(int fildes, off64_t offset, int whence);
+extern off_t      XrdPosix_Lseek(int fildes, off_t offset, int whence);
 
 extern int        XrdPosix_Lstat(const char *path, struct stat *buf);
 
@@ -153,12 +150,11 @@ extern DIR*       XrdPosix_Opendir(const char *path);
 
 extern long       XrdPosix_Pathconf(const char *path, int name);
 
-extern ssize_t  XrdPosix_Pread(int fildes, void *buf, size_t nbyte,
-                                 off64_t offset);
+extern ssize_t    XrdPosix_Pread(int fildes, void *buf, size_t nbyte, off_t offset);
 
-extern ssize_t  XrdPosix_Read(int fildes, void *buf, size_t nbyte);
+extern ssize_t    XrdPosix_Read(int fildes, void *buf, size_t nbyte);
 
-extern long long  XrdPosix_Readv(int fildes, const struct iovec *iov, int iovcnt);
+extern ssize_t    XrdPosix_Readv(int fildes, const struct iovec *iov, int iovcnt);
 
 extern struct dirent*   XrdPosix_Readdir  (DIR *dirp);
 extern struct dirent64* XrdPosix_Readdir64(DIR *dirp);
@@ -182,19 +178,17 @@ extern int        XrdPosix_Statfs(const char *path, struct statfs *buf);
 
 extern int        XrdPosix_Statvfs(const char *path, struct statvfs *buf);
 
-extern ssize_t  XrdPosix_Pwrite(int fildes, const void *buf,
-                                  size_t nbyte, off64_t offset);
+extern ssize_t    XrdPosix_Pwrite(int fildes, const void *buf, size_t nbyte, off_t offset);
 
 extern long       XrdPosix_Telldir(DIR *dirp);
 
-extern int        XrdPosix_Truncate(const char *path, off64_t offset);
+extern int        XrdPosix_Truncate(const char *path, off_t offset);
 
 extern int        XrdPosix_Unlink(const char *path);
 
-extern ssize_t  XrdPosix_Write(int fildes, const void *buf,
-                                 size_t nbyte);
+extern ssize_t    XrdPosix_Write(int fildes, const void *buf, size_t nbyte);
 
-extern ssize_t  XrdPosix_Writev(int fildes, const struct iovec *iov, int iovcnt);
+extern ssize_t    XrdPosix_Writev(int fildes, const struct iovec *iov, int iovcnt);
 
 #ifdef __cplusplus
 };
